build: fix 4 instances of -Wsign-compare
On Debian sid x32,

```
tests/lzxpress.cpp: In function 'int main(int, char**)':
tests/lzxpress.cpp:48:38: warning: comparison of integer expressions of different signedness: 'int' and 'std::size_t' {aka 'unsigned int'} [-Wsign-compare]
 if (ucomplen != std::size(b1))
exch/emsmdb/aux_ext.cpp: In function 'pack_result aux_ext_push_aux_info(EXT_PUSH*, const AUX_INFO&)':
exch/emsmdb/aux_ext.cpp:120:67: warning: comparison of integer expressions of different signedness: 'int' and 'uint32_t' {aka 'unsigned int'} [-Wsign-compare]
 if (compressed_len <= 0 || compressed_len >= subext.m_offset)
exch/emsmdb/rop_ext.cpp: In function 'pack_result rop_ext_make_rpc_ext(const void*, uint32_t, const ROP_BUFFER*, void*, uint32_t*)':
exch/emsmdb/rop_ext.cpp:2000:67: warning: comparison of integer expressions of different signedness: 'int' and 'uint32_ ' {aka 'unsigned int'} [-Wsign-compare]
 if (compressed_len <= 0 || compressed_len >= subext.m_offset)
exch/nsp/nsp_interface.cpp: In function 'ec_error_t nsp_interface_get_matches(NSPI_HANDLE, uint32_t, STAT*, const MID_ARRAY*, uint32_t, const NSPRES*, const NSP_PROPNAME*, uint32_t, MID_ARRAY**, const LPROPTAG_ARRAY*, NSP_ROWSET**)':
exch/nsp/nsp_interface.cpp:1304:100: warning: comparison of integer expressions of different signedness: 'gromox::ab_tree::ab_base::iterator::difference_type' {aka 'int'} and 'uint32_t' {aka 'unsigned int'} [-Wsign-compare]
 for (auto it = base->ubegin() + start_pos; it != base->uend() && it-base->ubegin() < total; ++it)
```

+++ exch/emsmdb/aux_ext.cpp
@@ -117,7 +117,7 @@ pack_result aux_ext_push_aux_info(EXT_PUSH *pext, const AUX_INFO &r) try
 			rpc_header_ext.flags &= ~RHE_FLAG_COMPRESSED;
 		} else {
 			auto compressed_len = lzxpress_compress(ext_buff.get(), subext.m_offset, tmp_buff.get(), ext_buff_size);
-			if (compressed_len <= 0 || compressed_len >= subext.m_offset) {
+			if (compressed_len <= 0 || static_cast<size_t>(compressed_len) >= subext.m_offset) {
 				/* if we can not get benefit from the
 					compression, unmask the compress bit */
 				rpc_header_ext.flags &= ~RHE_FLAG_COMPRESSED;
